refactor: custom hook 적용
