+++ src/containers/LunchApp.tsx
@@ -1,25 +1,20 @@
-import React, { useState } from 'react';
+import React from 'react';
 import RestaurantList from '../components/RestaurantList';
 import mockData from "../data/mockData.json";
-import { Category, Restaurant, Sort } from '../types/Restaurant';
+import { Restaurant } from '../types/Restaurant';
 import SelectBox from '../components/SelectBox';
 import Header from '../components/Header';
 import styled from 'styled-components';
 import Modal from '../components/RestaurantModal';
 import { category, sort } from '../data/selects';
+import useRestaurant from '../hooks/useRestaurants';
 
 function LunchApp() {
-  const [restaurants, setRestaurants] = useState<Restaurant[]>(mockData);
-  const [sortBy, setSortBy] = useState<Sort>('name');
-  const [categorizeBy, setCategorizeBy] = useState<Category>('all');
 
-  const updateSortBy = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    setSortBy(e.target.value as Sort);
-  }
-
-  const updateCategorizeBy = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    setCategorizeBy(e.target.value as Category);
-  }
+  const {
+    restaurants, sortBy, categorizeBy,
+    updateSortBy, updateCategorizeBy,
+  } = useRestaurant(mockData as Restaurant[]);
 
   return (
     <>
